api/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wifi-cinemas-api"
version = "0.1.0"
description = "WiFi Cinémas backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "firebase-admin>=6.3.0",
    "redis>=5.0.0",
    "websockets>=12.0",
    "httpx>=0.26.0",
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
    "python-multipart>=0.0.6",
    "qrcode[pil]>=7.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0","pytest-asyncio>=0.23.0","httpx>=0.26.0","ruff>=0.1.0"]
